TG-42: separated table into another component
